requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# LLM Providers
openai==1.6.1
anthropic==0.8.1
ollama==0.1.6

# LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Vector Databases
chromadb==0.4.22
weaviate-client==3.26.2
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Document Processing
pypdf==3.17.4
pdfplumber==0.10.3
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Evaluation
scikit-learn==1.3.2
nltk==3.8.1
rouge-score==0.1.2

# Visualization & UI
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# Utilities
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2

# Development
pytest==7.4.3
black==23.12.1
jupyter==1.0.0